Update to SolrNet 1.1.0

+++ SampleSolrApp/Models/SearchResultsModel.cs
@@ -18,7 +18,7 @@ namespace SampleSolrApp.Models
     public record SearchResultsModel(
         IReadOnlyList<Product> Results,
         IReadOnlyList<FacetModel> Facets,
-        int NumFound,
+        long NumFound,
         SearchParameters Parameters
     )
     {
